requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
nltk==3.6.2
oauthlib==3.1.1
PySocks==1.7.1
python-dotenv==0.19.0
pytz==2021.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
textblob==0.15.3
tqdm==4.61.2
tweepy==3.10.0
urllib3==1.26.6
Werkzeug==2.0.1
